requirements.txt
gradio>=4.0
openai>=1.0
networkx>=3.0
numpy>=1.24
pyyaml>=6.0
requests>=2.31
httpx>=0.25
aiohttp>=3.9
jieba>=0.42
rank-bm25>=0.2.2
sentence-transformers>=2.2
Pillow>=10.0
tqdm>=4.66
python-dotenv>=1.0
loguru>=0.7
chromadb>=0.4
